chore: update remark tools
Move from jest (which has experimental-only support for ESM) to tape.

BREAKING CHANGE: ESM only
